==> src/notebooks/03-feature-engineering.ipynb <==
# Feature Engineering

## Import des outils / jeu de données

import pandas as pd
import seaborn as sns

from src.config import data_folder
from src.utils import init_notebook

init_notebook()

df = pd.read_csv(
    f"{data_folder}/data-cleaned.csv", sep=",", index_col="ID", parse_dates=True
)

df.head()

## Création de variables

**Tableau.** Variables ajoutées

| :--- | :--- |
| **NbAcceptedCampaigns** | Nombre de campagnes marketing acceptées (hors variable cible) |
| **HasAcceptedCampaigns** | Le client a déjà accepté une campagne (1) ou non (0) |
| **NbChildren** | Nombre d'enfants au total (bas âge + adolescent) |

## todo : faire une var somme des achats sur le web
## todo : faire un client type par type d'achat (web / magasin) => est-ce que les gens qui achètent en web / magasin sont différents ?
## todo : créer une variable duréeClient qui dit depuis quand le client est inscrit

## todo: commenter la démarche

df["NbAcceptedCampaigns"] = 0

for i in range(1, 6):
    df["NbAcceptedCampaigns"] += df[f"AcceptedCmp{i}"].astype(int)

sns.histplot(df["NbAcceptedCampaigns"], discrete=True)

df["HasAcceptedCampaigns"] = df["NbAcceptedCampaigns"] > 0

sns.histplot(df["HasAcceptedCampaigns"].astype(str))

df_clients = df[df["HasAcceptedCampaigns"]]
df_not_clients = df[~df["HasAcceptedCampaigns"]]

## todo: vérifier si par exemple Kids et Teen apportent des informations différentes => si les deux sont complémentaires, on les laisse / sinon on les supprime au profit de notre nouvelle variable

df[["Kidhome", "Teenhome"]].value_counts().plot(kind="bar")

df["NbChildren"] = df["Kidhome"].astype(int) + df["Teenhome"].astype(int)

sns.histplot(df["NbChildren"])

## todo: créer une variable "NbAcceptedCampaignsWithResponse", qui contient donc la 6ème campagne marketing => l'objectif est de clusteriser les gens en fonction de "à quel point ils sont marketables" => c'est-à-dire, on présente nos résultats en disant sur qui il vaut mieux se concentrer pour faire des pubs

## todo: créer une variable "RevenuePerClient" => le revenu rapporté par chaque client => ce serait plutôt juste une variable type "KPI" à afficher dans un dashboard non ?

## todo: de même, créer une variable "RevenueWinePerClient" => le revenu de vin pour chaque client, et faire de même pour chaque produit => par la suite on pourra faire des représentations "quel client rapporte le plus de vin"

## Sauvegarde du Dataframe

df.to_csv(f"{data_folder}/data-cleaned-feature-engineering.csv")
